Split off scrolling logic into its own module
